Deduplicate AArch64 image logic
Dockerfile and build.yaml are now shared.

With regard to the Dockerfile changes:

gcc-x86_64-linux-gnu was not included before in amd64 (x86_64) images.
It is not needed as long as the additional prefix variables for gcc
(which are required for the arm64 image) are not set. If they are set,
then gcc will be searched for with its full prefixed name and not
found. Unfortunately there is no clean logic for conditionally running
ENV commands in Dockerfiles, so this was considered the cleanest choice.

Signed-off-by: Mike Beaton <mjsbeaton@gmail.com>